0) Updated to OpenSubdiv 3.5.0
This release of OpenSubdiv includes updates which improve build
configuration and cross platform build support including:
  - support for CMake 3.x
  - support for Python 3
  - suppression of additional compiler warnings
  - suppression of some GLSL shader compiler warnings

While there are API additions in this OpenSubdiv release, the USD
code base does not yet make use of any new API, so the source
remains compatible with OpenSubdiv 3.4.4.

This version of OpenSubdiv is not yet specified by the vfx reference
platform but there is an active proposal to update the platform
definition.