1) Updated USD build script for Apple OS
This adds support to build_usd.py for building native, cross architecture,
and universal binaries on macOS along with an option to enable code signing.

These aspects are controlled by two additional build arguments:
  --build-target {native,x86_64,arm64,universal}
  --codesign

Aspects of this facility are factored into a separate internal module
that is not intended for wider use and which may be removed or changed
significantly in future releases.

Contribution: Jon Creighton